requirements.txt
flask==2.3.0
psycopg2-binary==2.9.6
gunicorn==21.2.0